refactor(notebook): share notebook cell state stores
