+++ src/components/notebook/index.ts
@@ -100,6 +100,54 @@ export {
   type NavigateNotebookOutlineItemOptions,
 } from "./outline-navigation";
 export {
+  getCellById,
+  getCellIdsSnapshot,
+  getNotebookCellsSnapshot,
+  replaceNotebookCells,
+  resetNotebookCells,
+  subscribeIds,
+  updateCellById,
+  updateNotebookCells,
+  useCell,
+  useCellIds,
+  useMaterializeVersion,
+  useSourceVersion,
+  type NotebookCellMetadata,
+  type NotebookStoreCell,
+  type NotebookStoreCodeCell,
+  type NotebookStoreMarkdownCell,
+  type NotebookStoreOutput,
+  type NotebookStoreRawCell,
+} from "./state/cell-store";
+export {
+  flushCellUIState,
+  getFocusedCellId,
+  setExecutingCellIds,
+  setFocusedCellId,
+  setQueuedCellIds,
+  setSearchCurrentMatch,
+  setSearchQuery,
+  useCellQueuePriority,
+  useFocusedCellId,
+  useIsCellExecuting,
+  useIsCellFocused,
+  useIsCellQueued,
+  useIsGroupExecuting,
+  useIsNextCellFromFocused,
+  useIsPreviousCellFromFocused,
+  useNotebookCellUIStateBridge,
+  useSearchActiveOffset,
+  useSearchCurrentMatch,
+  useSearchQuery,
+  type NotebookCellUIStateBridgeInput,
+  type NotebookFindMatch,
+} from "./state/cell-ui-state";
+export {
+  createNotebookViewModelFromNotebookCells,
+  notebookCellToViewCell,
+  useNotebookViewModel,
+} from "./state/view-model-store";
+export {
   createNotebookViewModel,
   notebookViewCellsToOutlineItems,
   notebookViewCellsToReadOnlyCells,
